% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{province_population}
\alias{province_population}
\title{Total population segregated for every Canadian province.}
\usage{
province_population(provinceName = "Canada")
}
\arguments{
\item{provinceName}{a character/ string depicting the name of the province}
}
\value{
Data frame with all the columns for the entire summary for all the provinces individually would be
returned. The columns includes the date, province name, active_cases, active_cases_change, avaccine,
cases,cumulative_avaccine, cumulative_cases,cumulative_cvaccine, cumulative_deaths,   cumulative_dvaccine,
cumulative_recovered, cumulative_testing, cvaccine, deaths, dvaccine, recovered, testing and testing_info.
}
\description{
The total population count for every province is individual mentioned.
It processes the API and returns the json data corresponding to entered province.
If 'Canada' is passed as an argument the output displays the data for all the Canadian
provinces.The returned data is a data frame and,
}
\examples{
province_population("Ontario")
}
