% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{recovered}
\alias{recovered}
\title{Function for returning data frame for the Covid - 19 recovery rate in different provinces in Canada.}
\usage{
recovered(provinceName = "Canada")
}
\arguments{
\item{provinceName}{a character/ string depicting the name of the province}
}
\value{
Data frame for the recovered cases after COVID -19 corresponding to a particular province
}
\description{
Performed data wrangling and cleaning using the API for the Covid - 19 cases in Canada.
It processes the API and returns the data corresponding to one province
which is passed on as the argument. If user passes empty argument, so by default Canada is used which returns the data of whole Canada as a whole.
The returned data is a data frame and contains the columns including the recovery date, province, cumulative recovered, and count of recovery.
}
\examples{
recovered('Alberta')
}
