% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{summary_of_cases}
\alias{summary_of_cases}
\title{Covid-19 particular day\'s individual summary for every Canadian province}
\usage{
summary_of_cases(provinceName = "Canada")
}
\arguments{
\item{provinceName}{a character/ string depicting the name of the province}
}
\value{
Data frame with all the columns for the entire summary for all the provinces individually would be returned. The columns       #' includes the date, province name, active_cases, active_cases_change, avaccine, cases,cumulative_avaccine,
cumulative_cases,cumulative_cvaccine, cumulative_deaths,   cumulative_dvaccine, cumulative_recovered, cumulative_testing, cvaccine,
deaths, dvaccine, recovered, testing and testing_info.
}
\description{
Summary for all the provinces relating to Covid-19 active cases, avaccine, cvaccine and dvaccine. Moreover, is discloses   #' the count of people who have recovered and who were tested.
It processes the API and returns the json data corresponding to entered province. exceptional user input have been handled. If
Canada is passed as an argument the output displays the data for all the Canadian provinces.The returned data is a data frame and,
}
\examples{
summary_of_cases("Ontario")
}
