% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{summary_of_peaktime}
\alias{summary_of_peaktime}
\title{Summary of Covid-19 peak time which is 01 September 2020 for every Canadian province}
\usage{
summary_of_peaktime(provinceName = "Canada", date = "01-09-2020")
}
\arguments{
\item{provinceName}{a character/ string depicting the name of the province}

\item{date}{desired date - to get the summary on this date}
}
\value{
Data frame with all the columns for the entire summary for all the provinces individually would be returned. The columns
includes the date, province name, active_cases, active_cases_change, avaccine, cases,cumulative_avaccine,
cumulative_cases,cumulative_cvaccine, cumulative_deaths,   cumulative_dvaccine, cumulative_recovered, cumulative_testing, cvaccine,
deaths, dvaccine, recovered, testing and testing_info.
}
\description{
Summary for all the provinces relating to Covid-19 peak time in Canada which is the month of September 2020 or summary of
desired province on a desired date. It returns the
active cases, avaccine, cvaccine and dvaccine. Moreover, is discloses the count of people who have recovered and who were tested on
September 01, 2020.
It processes the API and returns the json data corresponding to entered province. If
'Canada' is passed as an argument the output displays the data for all the Canadian provinces.The returned data is a data frame and,
}
\examples{
summary_of_peaktime("Ontario")
}
