% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{yearly_deaths}
\alias{yearly_deaths}
\title{Function for returning data frame for the Covid - 19 deaths in different provinces in Canada according to the year.}
\usage{
yearly_deaths(provinceName = "Canada", dyear = "2020")
}
\arguments{
\item{provinceName}{a character/ string depicting the name of the province}

\item{dyear}{a character/ string showing the year}
}
\value{
Data frame for the Covid - 19 deaths corresponding to a particular province and year
}
\description{
Performed data wrangling and cleaning using the API for the Covid - 19 cases in Canada.
It processes the API and returns the data corresponding to one province
which is passed on as the argument. If user passes empty argument, so by default Canada is used which returns the data of whole Canada as a whole.
The returned data is a data frame and contains the columns including the date of the death, province name, cumulative deaths and deaths.
}
\examples{
yearly_deaths('Alberta','2020')
}
