% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{Yearlycases_as_per_province}
\alias{Yearlycases_as_per_province}
\title{Function for returning data frame for the Covid - 19 cases in different provinces as per the year specified in Canada.}
\usage{
Yearlycases_as_per_province(provinceName = "Canada", yearpassed = "2020")
}
\arguments{
\item{provinceName}{a character/ string depicting the name of the province}

\item{yearpassed}{a character/ string depicting the year}
}
\value{
Data frame for the Covid - 19 cases corresponding to a particular province and year
}
\description{
Performed data wrangling and cleaning using the API for the Covid - 19 cases in Canada.
It processes the API and returns the data corresponding to one province
which is passed on as the argument. If user passes empty argument, so by default Canada is used which returns the data of whole Canada as a whole.
Also, the data is returned as per the year. By default the year chosen is 2020.
The returned data is a data frame and contains the columns including the date of the cases, province name, cases and cumulative cases for the year 	    #' specified
}
\examples{
Yearlycases_as_per_province('Alberta','2020')
}
